+++ airflow-core/tests/unit/dags/test_assets.py
@@ -19,11 +19,11 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from airflow.exceptions import AirflowFailException, AirflowSkipException
 from airflow.models.dag import DAG
 from airflow.providers.standard.operators.bash import BashOperator
 from airflow.providers.standard.operators.python import PythonOperator
 from airflow.sdk.definitions.asset import Asset
+from airflow.sdk.exceptions import AirflowFailException, AirflowSkipException
 
 skip_task_dag_asset = Asset(uri="s3://dag_with_skip_task/output_1.txt", name="skip", extra={"hi": "bye"})
 fail_task_dag_asset = Asset(uri="s3://dag_with_fail_task/output_1.txt", name="fail", extra={"hi": "bye"})
