+++ providers/amazon/src/airflow/providers/amazon/aws/hooks/athena_sql.py
@@ -119,7 +119,7 @@ class AthenaSQLHook(AwsBaseHook, DbApiHook):
                     raise ValueError(
                         f"Encountered non-JSON in `extra` field for connection {self.aws_conn_id!r}."
                     )
-            except AirflowNotFoundException:
+            except (AirflowNotFoundException, RuntimeError):
                 connection = athena_conn
                 connection.conn_type = "aws"
                 self.log.warning(
