+++ providers/amazon/tests/unit/amazon/aws/operators/test_datasync.py
@@ -217,9 +217,11 @@ class TestDataSyncOperatorCreate(DataSyncTestCaseBase):
         with pytest.raises(AirflowException):
             self.set_up_operator(source_location_uri=None)
         with pytest.raises(AirflowException):
-            self.set_up_operator(destination_location_uri=None)
+            self.set_up_operator(destination_location_uri=None, task_id="datasync_task2")
         with pytest.raises(AirflowException):
-            self.set_up_operator(source_location_uri=None, destination_location_uri=None)
+            self.set_up_operator(
+                source_location_uri=None, destination_location_uri=None, task_id="datasync_task3"
+            )
         # ### Check mocks:
         mock_get_conn.assert_not_called()
 
@@ -441,9 +443,11 @@ class TestDataSyncOperatorGetTasks(DataSyncTestCaseBase):
         with pytest.raises(AirflowException):
             self.set_up_operator(source_location_uri=None)
         with pytest.raises(AirflowException):
-            self.set_up_operator(destination_location_uri=None)
+            self.set_up_operator(destination_location_uri=None, task_id="datasync_task2")
         with pytest.raises(AirflowException):
-            self.set_up_operator(source_location_uri=None, destination_location_uri=None)
+            self.set_up_operator(
+                source_location_uri=None, destination_location_uri=None, task_id="datasync_task3"
+            )
         # ### Check mocks:
         mock_get_conn.assert_not_called()
