+++ providers/cncf/kubernetes/src/airflow/providers/cncf/kubernetes/version_compat.py
@@ -35,4 +35,15 @@ def get_base_airflow_version_tuple() -> tuple[int, int, int]:
 AIRFLOW_V_3_0_PLUS = get_base_airflow_version_tuple() >= (3, 0, 0)
 AIRFLOW_V_3_1_PLUS = get_base_airflow_version_tuple() >= (3, 1, 0)
 
-__all__ = ["AIRFLOW_V_3_0_PLUS", "AIRFLOW_V_3_1_PLUS"]
+try:
+    from airflow.sdk.exceptions import AirflowSkipException, TaskDeferred
+except ImportError:
+    from airflow.exceptions import AirflowSkipException, TaskDeferred  # type: ignore[no-redef, attr-defined]
+
+
+__all__ = [
+    "AIRFLOW_V_3_0_PLUS",
+    "AIRFLOW_V_3_1_PLUS",
+    "AirflowSkipException",
+    "TaskDeferred",
+]
