+++ providers/cncf/kubernetes/tests/unit/cncf/kubernetes/decorators/test_kubernetes_cmd.py
@@ -20,7 +20,7 @@ import contextlib
 
 import pytest
 
-from airflow.exceptions import AirflowSkipException
+from airflow.providers.cncf.kubernetes.version_compat import AirflowSkipException
 
 from tests_common.test_utils.version_compat import AIRFLOW_V_3_0_PLUS
