+++ providers/common/sql/tests/unit/common/sql/operators/test_sql.py
@@ -59,7 +59,11 @@ from tests_common.test_utils.dag import sync_dag_to_db
 from tests_common.test_utils.db import clear_db_dag_bundles, clear_db_dags, clear_db_runs, clear_db_xcom
 from tests_common.test_utils.markers import skip_if_force_lowest_dependencies_marker
 from tests_common.test_utils.providers import get_provider_min_airflow_version
-from tests_common.test_utils.version_compat import AIRFLOW_V_3_0_1, AIRFLOW_V_3_0_PLUS
+from tests_common.test_utils.version_compat import (
+    AIRFLOW_V_3_0_1,
+    AIRFLOW_V_3_0_PLUS,
+    AirflowFailException,
+)
 
 if AIRFLOW_V_3_0_PLUS:
     from airflow.utils.types import DagRunTriggeredByType
@@ -867,7 +871,7 @@ class TestIntervalCheckOperator:
         )
 
     def test_invalid_ratio_formula(self):
-        with pytest.raises(AirflowException, match="Invalid diff_method"):
+        with pytest.raises(AirflowFailException, match="Invalid diff_method"):
             self._construct_operator(
                 table="test_table",
                 metric_thresholds={
