+++ providers/google/tests/unit/google/cloud/sensors/test_pubsub.py
@@ -28,6 +28,8 @@ from airflow.exceptions import AirflowException, TaskDeferred
 from airflow.providers.google.cloud.sensors.pubsub import PubSubPullSensor
 from airflow.providers.google.cloud.triggers.pubsub import PubsubPullTrigger
 
+from tests_common.test_utils.version_compat import AirflowSensorTimeout
+
 TASK_ID = "test-task-id"
 TEST_PROJECT = "test-project"
 TEST_SUBSCRIPTION = "test-subscription"
@@ -111,7 +113,7 @@ class TestPubSubPullSensor:
 
         mock_hook.return_value.pull.return_value = []
 
-        with pytest.raises(AirflowException):
+        with pytest.raises(AirflowSensorTimeout):
             operator.execute({})
 
     @mock.patch("airflow.providers.google.cloud.sensors.pubsub.PubSubHook")
