+++ providers/samba/tests/unit/samba/hooks/test_samba.py
@@ -41,7 +41,7 @@ PATH_PARAMETER_NAMES = {"path", "src", "dst"}
 class TestSambaHook:
     @pytest.mark.db_test
     def test_get_conn_should_fail_if_conn_id_does_not_exist(self, sdk_connection_not_found):
-        with pytest.raises(AirflowNotFoundException):
+        with pytest.raises((AirflowNotFoundException, RuntimeError)):
             SambaHook("non-existed-connection-id")
 
     @mock.patch("smbclient.register_session")
