+++ providers/telegram/tests/unit/telegram/hooks/test_telegram.py
@@ -72,7 +72,7 @@ class TestTelegramHook:
 
     @pytest.mark.db_test
     def test_should_raise_exception_if_conn_id_doesnt_exist(self, sdk_connection_not_found):
-        with pytest.raises(airflow.exceptions.AirflowNotFoundException) as ctx:
+        with pytest.raises((airflow.exceptions.AirflowNotFoundException, RuntimeError)) as ctx:
             TelegramHook(telegram_conn_id="telegram-webhook-non-existent")
 
         assert str(ctx.value) == "The conn_id `telegram-webhook-non-existent` isn't defined"
