+++ task-sdk/src/airflow/sdk/bases/sensor.py
@@ -25,7 +25,10 @@ from collections.abc import Callable, Iterable
 from datetime import timedelta
 from typing import TYPE_CHECKING, Any
 
-from airflow.exceptions import (
+from airflow.sdk.configuration import conf
+from airflow.sdk import timezone
+from airflow.sdk.bases.operator import BaseOperator
+from airflow.sdk.exceptions import (
     AirflowException,
     AirflowFailException,
     AirflowRescheduleException,
@@ -35,9 +38,6 @@ from airflow.exceptions import (
     TaskDeferralError,
     TaskDeferralTimeout,
 )
-from airflow.sdk import timezone
-from airflow.sdk.bases.operator import BaseOperator
-from airflow.sdk.configuration import conf
 
 if TYPE_CHECKING:
     from airflow.sdk.definitions.context import Context
@@ -145,9 +145,7 @@ class BaseSensorOperator(BaseOperator):
             return poke_interval
         if isinstance(poke_interval, (int, float)) and poke_interval >= 0:
             return timedelta(seconds=poke_interval)
-        raise AirflowException(
-            "Operator arg `poke_interval` must be timedelta object or a non-negative number"
-        )
+        raise RuntimeError("Operator arg `poke_interval` must be timedelta object or a non-negative number")
 
     @staticmethod
     def _coerce_timeout(timeout: float | timedelta) -> timedelta:
@@ -155,7 +153,7 @@ class BaseSensorOperator(BaseOperator):
             return timeout
         if isinstance(timeout, (int, float)) and timeout >= 0:
             return timedelta(seconds=timeout)
-        raise AirflowException("Operator arg `timeout` must be timedelta object or a non-negative number")
+        raise RuntimeError("Operator arg `timeout` must be timedelta object or a non-negative number")
 
     @staticmethod
     def _coerce_max_wait(max_wait: float | timedelta | None) -> timedelta | None:
@@ -163,22 +161,22 @@ class BaseSensorOperator(BaseOperator):
             return max_wait
         if isinstance(max_wait, (int, float)) and max_wait >= 0:
             return timedelta(seconds=max_wait)
-        raise AirflowException("Operator arg `max_wait` must be timedelta object or a non-negative number")
+        raise RuntimeError("Operator arg `max_wait` must be timedelta object or a non-negative number")
 
     def _validate_input_values(self) -> None:
         if not isinstance(self.poke_interval, (int, float)) or self.poke_interval < 0:
-            raise AirflowException("The poke_interval must be a non-negative number")
+            raise RuntimeError("The poke_interval must be a non-negative number")
         if not isinstance(self.timeout, (int, float)) or self.timeout < 0:
-            raise AirflowException("The timeout must be a non-negative number")
+            raise RuntimeError("The timeout must be a non-negative number")
         if self.mode not in self.valid_modes:
-            raise AirflowException(
+            raise RuntimeError(
                 f"The mode must be one of {self.valid_modes},'{self.dag.dag_id if self.has_dag() else ''} "
                 f".{self.task_id}'; received '{self.mode}'."
             )
 
     def poke(self, context: Context) -> bool | PokeReturnValue:
         """Override when deriving this class."""
-        raise AirflowException("Override me.")
+        raise RuntimeError("Override me.")
 
     def execute(self, context: Context) -> Any:
         started_at: datetime.datetime | float
@@ -255,7 +253,7 @@ class BaseSensorOperator(BaseOperator):
             return super().resume_execution(next_method, next_kwargs, context)
         except TaskDeferralTimeout as e:
             raise AirflowSensorTimeout(*e.args) from e
-        except (AirflowException, TaskDeferralError) as e:
+        except (RuntimeError, AirflowException, TaskDeferralError) as e:
             if self.soft_fail:
                 raise AirflowSkipException(str(e)) from e
             raise
