+++ task-sdk/src/airflow/sdk/definitions/decorators/setup_teardown.py
@@ -20,7 +20,6 @@ import types
 from collections.abc import Callable
 from typing import TYPE_CHECKING, cast
 
-from airflow.exceptions import AirflowException
 from airflow.sdk.bases.operator import BaseOperator
 from airflow.sdk.definitions._internal.setup_teardown import SetupTeardownContext
 from airflow.sdk.definitions.decorators.task_group import _TaskGroupFactory
@@ -52,7 +51,7 @@ def setup_task(func: Callable) -> Callable:
     if isinstance(func, types.FunctionType):
         func = python_task(func)
     if isinstance(func, _TaskGroupFactory):
-        raise AirflowException("Task groups cannot be marked as setup or teardown.")
+        raise RuntimeError("Task groups cannot be marked as setup or teardown.")
     func = cast("_TaskDecorator", func)
     func.is_setup = True
     return func
@@ -77,7 +76,7 @@ def teardown_task(_func=None, *, on_failure_fail_dagrun: bool = False) -> Callab
         if isinstance(func, types.FunctionType):
             func = python_task(func)
         if isinstance(func, _TaskGroupFactory):
-            raise AirflowException("Task groups cannot be marked as setup or teardown.")
+            raise RuntimeError("Task groups cannot be marked as setup or teardown.")
         func = cast("_TaskDecorator", func)
 
         func.is_teardown = True
@@ -102,9 +101,9 @@ class ContextWrapper(list):
             if isinstance(task, BaseOperator):
                 operators.append(task)
                 if not task.is_setup and not task.is_teardown:
-                    raise AirflowException("Only setup/teardown tasks can be used as context managers.")
+                    raise RuntimeError("Only setup/teardown tasks can be used as context managers.")
             elif not task.operator.is_setup and not task.operator.is_teardown:
-                raise AirflowException("Only setup/teardown tasks can be used as context managers.")
+                raise RuntimeError("Only setup/teardown tasks can be used as context managers.")
         if not operators:
             # means we have XComArgs
             operators = [task.operator for task in self.tasks]
