+++ task-sdk/tests/task_sdk/bases/test_operator.py
@@ -807,7 +807,7 @@ class TestBaseOperator:
             pass
 
         # The following throws an exception if metaclass breaks MRO:
-        #   airflow.exceptions.AirflowException: Invalid arguments were passed to Branch (task_id: test). Invalid arguments were:
+        #   airflow.sdk.exceptions.AirflowException: Invalid arguments were passed to Branch (task_id: test). Invalid arguments were:
         #   **kwargs: {'sql': 'sql', 'follow_task_ids_if_true': ['x'], 'follow_task_ids_if_false': ['y']}
         op = Branch(
             task_id="test",
