+++ task-sdk/tests/task_sdk/bases/test_sensor.py
@@ -24,19 +24,18 @@ from unittest.mock import Mock
 import pytest
 import time_machine
 
-from airflow.exceptions import (
-    AirflowException,
+from airflow.models.trigger import TriggerFailureReason
+from airflow.providers.standard.operators.empty import EmptyOperator
+from airflow.sdk import TaskInstanceState, timezone
+from airflow.sdk.bases.sensor import BaseSensorOperator, PokeReturnValue, poke_mode_only
+from airflow.sdk.definitions.dag import DAG
+from airflow.sdk.exceptions import (
     AirflowFailException,
     AirflowRescheduleException,
     AirflowSensorTimeout,
     AirflowSkipException,
     AirflowTaskTimeout,
 )
-from airflow.models.trigger import TriggerFailureReason
-from airflow.providers.standard.operators.empty import EmptyOperator
-from airflow.sdk import TaskInstanceState, timezone
-from airflow.sdk.bases.sensor import BaseSensorOperator, PokeReturnValue, poke_mode_only
-from airflow.sdk.definitions.dag import DAG
 from airflow.sdk.execution_time.comms import RescheduleTask, TaskRescheduleStartDate
 from airflow.sdk.timezone import datetime
 
@@ -64,7 +63,7 @@ class DummyAsyncSensor(BaseSensorOperator):
         self.return_value = return_value
 
     def execute_complete(self, context, event=None):
-        raise AirflowException("Should be skipped")
+        raise RuntimeError("Should be skipped")
 
 
 class DummySensorWithXcomValue(BaseSensorOperator):
@@ -274,7 +273,7 @@ class TestBaseSensor:
         assert state == TaskInstanceState.SUCCESS
 
     def test_invalid_mode(self):
-        with pytest.raises(AirflowException):
+        with pytest.raises(RuntimeError):
             DummySensor(task_id="a", mode="foo")
 
     def test_ok_with_custom_reschedule_exception(self, make_sensor, run_task):
@@ -311,7 +310,7 @@ class TestBaseSensor:
         negative_poke_interval = -10
         non_number_poke_interval = "abcd"
         positive_poke_interval = 10
-        with pytest.raises(AirflowException):
+        with pytest.raises(RuntimeError):
             DummySensor(
                 task_id="test_sensor_task_1",
                 return_value=None,
@@ -319,7 +318,7 @@ class TestBaseSensor:
                 timeout=25,
             )
 
-        with pytest.raises(AirflowException):
+        with pytest.raises(RuntimeError):
             DummySensor(
                 task_id="test_sensor_task_2",
                 return_value=None,
@@ -335,12 +334,12 @@ class TestBaseSensor:
         negative_timeout = -25
         non_number_timeout = "abcd"
         positive_timeout = 25
-        with pytest.raises(AirflowException):
+        with pytest.raises(RuntimeError):
             DummySensor(
                 task_id="test_sensor_task_1", return_value=None, poke_interval=10, timeout=negative_timeout
             )
 
-        with pytest.raises(AirflowException):
+        with pytest.raises(RuntimeError):
             DummySensor(
                 task_id="test_sensor_task_2", return_value=None, poke_interval=10, timeout=non_number_timeout
             )
@@ -669,7 +668,7 @@ class TestAsyncSensor:
         ("soft_fail", "expected_exception"),
         [
             (True, AirflowSkipException),
-            (False, AirflowException),
+            (False, RuntimeError),
         ],
     )
     def test_fail_after_resuming_deferred_sensor(self, soft_fail, expected_exception):
