+++ task-sdk/tests/task_sdk/definitions/decorators/test_setup_teardown.py
@@ -19,7 +19,6 @@ from __future__ import annotations
 
 import pytest
 
-from airflow.exceptions import AirflowException
 from airflow.providers.standard.operators.bash import BashOperator
 from airflow.sdk import DAG, setup, task, task_group, teardown
 from airflow.sdk.definitions.decorators.setup_teardown import context_wrapper
@@ -124,7 +123,7 @@ class TestSetupTearDownTask:
     def test_setup_taskgroup_decorator(self):
         with DAG("test_setup_taskgroup_decorator"):
             with pytest.raises(  # noqa: PT012, check decorators required more than one line
-                expected_exception=AirflowException,
+                expected_exception=RuntimeError,
                 match="Task groups cannot be marked as setup or teardown.",
             ):
 
@@ -142,7 +141,7 @@ class TestSetupTearDownTask:
     def test_teardown_taskgroup_decorator(self):
         with DAG("test_teardown_taskgroup_decorator"):
             with pytest.raises(  # noqa: PT012, check decorators required more than one line
-                expected_exception=AirflowException,
+                expected_exception=RuntimeError,
                 match="Task groups cannot be marked as setup or teardown.",
             ):
 
@@ -367,7 +366,7 @@ class TestSetupTearDownTask:
 
         with DAG("test_setup_teardown_as_context_manager_normal_tasks_rel_set_up"):
             with pytest.raises(
-                AirflowException, match="Only setup/teardown tasks can be used as context managers."
+                RuntimeError, match="Only setup/teardown tasks can be used as context managers."
             ):
                 with mytask():
                     pass
